requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
python-dotenv